GLOBAL: Integrate TI-NSPIRE port (#90)
* nspire: Add initial

* NSPIRE: Remove -Werror and -fanalyzer

* .gitignore: Modify to exclude .tns files

* Add Tyler's fixes for latest SDK

* Remove yml

* nspire: Render zombie further than alias model

* Re-add github actions

* NSPIRE: port restructure from upstream vril

* Update YAML to add TNS build

* NSPIRE: Fix blood printing

* Support compiling with -Werror

* Remove fixed div for non nspire

* Revise makefile build name and actions

* Update README.md

* Add large texture gating

* Update README.md

* Revert condebug hack

* COMMON: Revise NSPIRE specific blood

* Migrate more NSPIRE specifics to common with stubs

* NSPIRE: Limit ifdef in nzportable_def

* TNS: Add Vril features to clean up Nspire-specific common changes

---------

+++ source/pr_edict.c
@@ -1044,7 +1044,7 @@ void PR_LoadProgs (void)
 		CRC_ProcessByte (&pr_crc, ((byte *)progs)[i]);
 
 // byte swap the header
-	for (i=0 ; i<sizeof(*progs)/4 ; i++)
+	for (i=0 ; i < (int)sizeof(*progs)/4 ; i++)
 		((int *)progs)[i] = LittleLong ( ((int *)progs)[i] );
 
 	pr_functions = (dfunction_t *)((byte *)progs + progs->ofs_functions);
